Moved endpoints-related tests into new package org.ga4gh.cts.api.endpoints, new class RawEndpointsIT. Fix for #51.
