+++ cts-java/src/test/java/org/ga4gh/cts/api/endpoints/EndpointsTests.java
@@ -1,0 +1,7 @@
+package org.ga4gh.cts.api.endpoints;
+
+/**
+ * Marker interface for tests related to endpoints.
+ */
+public interface EndpointsTests { /* category marker */
+}
